# openapi.yaml
openapi: 3.0.0
info:
  title: OrcaSlicer CLI Headless Runner API
  description: A complete RESTful API wrapping OrcaSlicer CLI to perform 3D model slicing (supporting multi-filament setups) and automatic plate arrangement/orientation (3MF files).
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local Docker container server
paths:
  /api/slice/start:
    post:
      summary: Start a new slicing job (RESTful)
      description: Uploads a raw 3D model file (STL or 3MF) and a combined JSON configuration object. Slicing runs asynchronously in the background.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The raw 3D mesh model file (.stl or .3mf).
                config:
                  type: string
                  description: A JSON-string representing the combined slicing configuration parameters.
                  example: |
                    {
                      "printer": "default/machine/Creality Ender-3 0.4 nozzle.json",
                      "process": "default/process/0.16mm Optimal @Creality Ender3 0.4.json",
                      "plate": 0,
                      "filaments": {
                        "1": "default/filament/Creality Generic PLA.json",
                        "2": "default/filament/Creality Generic PETG.json"
                      }
                    }
            encoding:
              config:
                contentType: application/json
      responses:
        '200':
          description: Slicing job initialized successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    format: uuid
                    example: f30b6adf-e1fa-4d74-bc82-f61244179fac
                  status:
                    type: string
                    example: pending
                  message:
                    type: string
                    example: Slicing job started in background.
        '400':
          description: Invalid request parameters or malformed configuration JSON.
  /api/slice/status/{job_id}:
    get:
      summary: Retrieve slicing job status
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier of the slicing job.
      responses:
        '200':
          description: Job status fetched.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [pending, slicing, completed, failed]
                  sliced_file:
                    type: string
                    nullable: true
                    example: plate_1.gcode
                  error:
                    type: string
                    nullable: true
        '404':
          description: Job not found.
  /api/slice/logs/{job_id}:
    get:
      summary: Stream real-time slicing stdout/stderr console logs
      description: Streams OrcaSlicer's output logs line-by-line using Server-Sent Events (SSE). Keep-alive connection remains open until slicing finishes.
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: SSE event stream of log strings.
          content:
            text/event-stream:
              schema:
                type: string
                example: "data: Loading printer profile: Creality Ender-3\n\n"
  /api/slice/download/{job_id}:
    get:
      summary: Download the sliced G-code or 3MF file
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File payload stream.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Job is not complete or failed.
        '404':
          description: File or job not found.
  /api/arrange:
    post:
      summary: Auto-arrange and orient a 3MF project file
      description: Uploads a 3MF file, arranges and/or orients the models inside it using OrcaSlicer's layout solver, and synchronously streams the updated 3MF file back.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The raw .3mf project file to be arranged.
                arrange:
                  type: boolean
                  default: true
                  description: Whether to auto-arrange layout.
                orient:
                  type: boolean
                  default: true
                  description: Whether to auto-orient parts to minimize overhangs.
      responses:
        '200':
          description: Succeeded. Streams the arranged .3mf binary payload directly.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Slicer process failed or input file was invalid.
  /api/profiles:
    get:
      summary: List loaded presets and profiles
      responses:
        '200':
          description: Presets scanned successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  machine:
                    type: array
                    items:
                      $ref: '#/components/schemas/Profile'
                  process:
                    type: array
                    items:
                      $ref: '#/components/schemas/Profile'
                  filament:
                    type: array
                    items:
                      $ref: '#/components/schemas/Profile'
  /api/profiles/upload:
    post:
      summary: Upload a preset profile JSON directly to the config volume
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [machine, process, filament]
                  description: The preset target profile folder.
                file:
                  type: string
                  format: binary
                  description: The OrcaSlicer preset JSON file.
      responses:
        '200':
          description: Profile uploaded.
        '400':
          description: Invalid file format or target directory.
  /api/health:
    get:
      summary: Check container and OrcaSlicer service health
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  orcaslicer_installed:
                    type: boolean
                    example: true
                  config_mounted:
                    type: boolean
                    example: true
components:
  schemas:
    SliceConfig:
      type: object
      required:
        - printer
        - process
        - filaments
      properties:
        printer:
          type: string
          description: Path or name of the printer preset JSON (scannable from `/api/profiles`).
        process:
          type: string
          description: Path or name of the process preset JSON (scannable from `/api/profiles`).
        plate:
          type: integer
          default: 0
          description: Target build plate index to slice (0 slices all plates).
        filaments:
          type: object
          additionalProperties:
            type: string
          description: Mapping of extruder slots (1-indexed keys like "1", "2") to filament preset JSON paths.
    Profile:
      type: object
      properties:
        name:
          type: string
          example: "Creality Ender-3 0.4 nozzle"
        filename:
          type: string
          example: "Creality Ender-3 0.4 nozzle.json"
        rel_path:
          type: string
          example: "default/machine/Creality Ender-3 0.4 nozzle.json"
        full_path:
          type: string
          example: "/config/user/default/machine/Creality Ender-3 0.4 nozzle.json"